+++ server/src/scanner/mod.rs
@@ -29,5 +29,6 @@ pub mod sprite_scanner;
 pub mod state_category_scanner;
 pub mod state_scanner;
 pub mod strategic_region_scanner;
+pub mod terrain_scanner;
 pub mod trait_scanner;
 pub mod variable_scanner;